slabs: revert downloads_test.go
